requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
multidict==4.7.5
python-aiml==0.9.3
python-dotenv==0.14.0
requests==2.23.0
rfc3986==1.4.0
sniffio==1.1.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
